requirements.txt
aws-cdk-lib==2.86.0
aws_cdk.aws_lambda_python_alpha==2.86.0a0
constructs>=10.0.0,<11.0.0
aws_solutions_constructs.aws-apigateway-lambda==2.39.0
